abstracted out topics routes
